--- a/pkg/astvalidation/definition_validation_test.go
+++ b/pkg/astvalidation/definition_validation_test.go
@@ -1,6 +1,10 @@
 package astvalidation
 
 import (
+	"fmt"
+	"io"
+	"os"
+	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -10,7 +14,28 @@ import (
 	"github.com/TykTechnologies/graphql-go-tools/pkg/asttransform"
 )
 
-func runDefinitionValidation(t *testing.T, definitionInput string, expectation ValidationState, rules ...Rule) {
+func TestMain(m *testing.M) {
+	printMemUsage := func(out io.Writer) {
+		bToMb := func(b uint64) uint64 {
+			return b / 1024 / 1024
+		}
+		var m runtime.MemStats
+		runtime.ReadMemStats(&m)
+
+		// For info on each, see: https://golang.org/pkg/runtime/#MemStats
+		fmt.Fprintf(out, "Alloc = %v MiB", bToMb(m.Alloc))
+		fmt.Fprintf(out, "\tTotalAlloc = %v MiB", bToMb(m.TotalAlloc))
+		fmt.Fprintf(out, "\tSys = %v MiB", bToMb(m.Sys))
+		fmt.Fprintf(out, "\tNumGC = %v\n", m.NumGC)
+	}
+
+	printMemUsage(os.Stderr)
+	exitCode := m.Run()
+	printMemUsage(os.Stderr)
+	os.Exit(exitCode)
+}
+
+func runDefinitionValidation(t testing.TB, definitionInput string, expectation ValidationState, rules ...Rule) {
 	definition, report := astparser.ParseGraphqlDocumentString(definitionInput)
 	require.False(t, report.HasErrors())