@titpetric 2025-02-15
Closed, implemented with #431